move dashed legend to helpers
