+++ src/charts/overlap_bar/render.js
@@ -9,7 +9,7 @@ import {
   setupChartArea,
 } from '../../utils/helpers/commonChartHelpers'
 import { preventOverflow, toClassText } from '../../utils/helpers/general'
-import { dashedLegend } from './dashedLegend'
+import { dashedLegend } from '../../utils/helpers/dashedLegend'
 
 export function renderChart({
   data,
